# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Velero API Server
  description: |
    REST API wrapper for Velero backup and restore operations with multi-cluster support.

    ## Features
    - Multi-cluster management
    - Backup and restore operations
    - Scheduled backups
    - Backup storage location management

    ## Multi-Cluster Support
    All API endpoints (except `/api/v1/clusters`) support cluster selection via:
    - Query parameter: `?cluster=cluster-name`
    - HTTP header: `X-Cluster: cluster-name`

    If not specified, the default cluster (first loaded) will be used.

  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /healthz:
    get:
      summary: Health check
      tags:
        - Health
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/clusters:
    get:
      summary: List available clusters
      tags:
        - Cluster
      description: Returns all configured Kubernetes clusters
      responses:
        '200':
          description: List of clusters
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          clusters:
                            type: array
                            items:
                              type: string
                            example:
                              - config-fat01
                              - config-fat02
                          default:
                            type: string
                            example: config-fat01

  /api/v1/backups:
    post:
      summary: Create a new backup
      tags:
        - Backup
      parameters:
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBackupRequest'
            example:
              name: backup-default-all
              includedNamespaces:
                - default
              includedResources:
                - deployments.apps
                - services
              ttl: "720h"
      responses:
        '201':
          description: Backup created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BackupResponse'
        '400':
          description: Invalid request or cluster not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

    get:
      summary: List all backups
      tags:
        - Backup
      parameters:
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: List of backups
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BackupResponse'

  /api/v1/backups/{name}:
    get:
      summary: Get backup details
      tags:
        - Backup
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Backup name
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: Backup details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BackupResponse'

    delete:
      summary: Delete a backup
      tags:
        - Backup
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Backup name
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: Backup deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/restores:
    post:
      summary: Create a new restore
      tags:
        - Restore
      parameters:
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRestoreRequest'
            example:
              name: restore-default-20260625
              backupName: backup-default-all
              includedNamespaces:
                - default
      responses:
        '201':
          description: Restore created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RestoreResponse'

    get:
      summary: List all restores
      tags:
        - Restore
      parameters:
        - name: backupName
          in: query
          required: false
          schema:
            type: string
          description: Filter by backup name
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: List of restores
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/RestoreResponse'

  /api/v1/restores/{name}:
    get:
      summary: Get restore details
      tags:
        - Restore
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Restore name
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: Restore details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RestoreResponse'

  /api/v1/schedules:
    post:
      summary: Create a backup schedule
      tags:
        - Schedule
      parameters:
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
            example:
              name: daily-backup
              schedule: "0 2 * * *"
              includedNamespaces:
                - default
              ttl: "720h"
      responses:
        '201':
          description: Schedule created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ScheduleResponse'

    get:
      summary: List all schedules
      tags:
        - Schedule
      parameters:
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: List of schedules
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ScheduleResponse'

  /api/v1/schedules/{name}:
    get:
      summary: Get schedule details
      tags:
        - Schedule
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Schedule name
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: Schedule details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ScheduleResponse'

    delete:
      summary: Delete a schedule
      tags:
        - Schedule
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Schedule name
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: Schedule deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'

  /api/v1/storage-locations:
    get:
      summary: List backup storage locations
      tags:
        - Storage Location
      parameters:
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: List of storage locations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/BSLResponse'

  /api/v1/storage-locations/{name}:
    get:
      summary: Get storage location details
      tags:
        - Storage Location
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Storage location name
        - $ref: '#/components/parameters/ClusterQueryParam'
        - $ref: '#/components/parameters/ClusterHeaderParam'
      responses:
        '200':
          description: Storage location details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/APIResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BSLResponse'

components:
  parameters:
    ClusterQueryParam:
      name: cluster
      in: query
      required: false
      schema:
        type: string
      description: Target cluster name (uses default if not specified)
      example: config-fat02

    ClusterHeaderParam:
      name: X-Cluster
      in: header
      required: false
      schema:
        type: string
      description: Target cluster name via header (query param takes precedence)
      example: config-fat02

  schemas:
    APIResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          nullable: true

    CreateBackupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Backup name
          example: backup-default-all
        includedNamespaces:
          type: array
          items:
            type: string
          description: Namespaces to include (empty means all namespaces)
          example:
            - default
            - kube-system
        excludedNamespaces:
          type: array
          items:
            type: string
          description: Namespaces to exclude from backup
        includedResources:
          type: array
          items:
            type: string
          description: Resource types to include (empty means all resources)
          example:
            - deployments.apps
            - services
            - configmaps
        excludedResources:
          type: array
          items:
            type: string
          description: Resource types to exclude
        labelSelector:
          type: object
          additionalProperties:
            type: string
          description: Label selector for filtering resources
          example:
            app: myapp
        storageLocation:
          type: string
          description: Backup storage location name (uses default if not specified)
          example: default
        ttl:
          type: string
          description: Backup retention time (e.g., "720h" for 30 days)
          example: "720h"
        snapshotVolumes:
          type: boolean
          description: Whether to snapshot volumes
          example: false

    BackupResponse:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        phase:
          type: string
          description: Backup phase (New, InProgress, Completed, Failed, etc.)
          enum:
            - New
            - InProgress
            - Completed
            - Failed
            - PartiallyFailed
        includedNamespaces:
          type: array
          items:
            type: string
        excludedNamespaces:
          type: array
          items:
            type: string
        storageLocation:
          type: string
        ttl:
          type: string
        startTimestamp:
          type: string
          format: date-time
          nullable: true
        completionTimestamp:
          type: string
          format: date-time
          nullable: true
        errors:
          type: integer
        warnings:
          type: integer
        labels:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time

    CreateRestoreRequest:
      type: object
      required:
        - name
        - backupName
      properties:
        name:
          type: string
          description: Restore name
          example: restore-default-20260625
        backupName:
          type: string
          description: Name of the backup to restore from
          example: backup-default-all
        includedNamespaces:
          type: array
          items:
            type: string
          description: Namespaces to restore (empty means restore to original namespaces)
          example:
            - default
        excludedNamespaces:
          type: array
          items:
            type: string
          description: Namespaces to exclude from restore
        includedResources:
          type: array
          items:
            type: string
          description: Resource types to restore (empty means all resources from backup)
          example:
            - deployments.apps
            - services
        excludedResources:
          type: array
          items:
            type: string
          description: Resource types to exclude from restore
        labelSelector:
          type: object
          additionalProperties:
            type: string
          description: Label selector for filtering resources
        restorePVs:
          type: boolean
          description: Whether to restore persistent volumes
          example: false

    RestoreResponse:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        backupName:
          type: string
        phase:
          type: string
          description: Restore phase
          enum:
            - New
            - InProgress
            - Completed
            - Failed
            - PartiallyFailed
        startTimestamp:
          type: string
          format: date-time
          nullable: true
        completionTimestamp:
          type: string
          format: date-time
          nullable: true
        errors:
          type: integer
        warnings:
          type: integer
        createdAt:
          type: string
          format: date-time

    CreateScheduleRequest:
      type: object
      required:
        - name
        - schedule
      properties:
        name:
          type: string
          description: Schedule name
          example: daily-backup
        schedule:
          type: string
          description: Cron expression (minute hour day month weekday)
          example: "0 2 * * *"
        includedNamespaces:
          type: array
          items:
            type: string
          description: Namespaces to backup
          example:
            - default
        excludedNamespaces:
          type: array
          items:
            type: string
        includedResources:
          type: array
          items:
            type: string
          description: Resource types to backup
        excludedResources:
          type: array
          items:
            type: string
        labelSelector:
          type: object
          additionalProperties:
            type: string
        storageLocation:
          type: string
        ttl:
          type: string
          example: "720h"

    ScheduleResponse:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        schedule:
          type: string
        phase:
          type: string
          enum:
            - Enabled
            - FailedValidation
        lastBackup:
          type: string
          format: date-time
          nullable: true
        storageLocation:
          type: string
        ttl:
          type: string
        createdAt:
          type: string
          format: date-time

    BSLResponse:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        provider:
          type: string
          description: Storage provider
          example: aws
          enum:
            - aws
            - gcp
            - azure
            - minio
        bucket:
          type: string
          example: velero-backups
        prefix:
          type: string
          example: backups
        phase:
          type: string
          enum:
            - Available
            - Unavailable
        lastSyncedTime:
          type: string
          format: date-time
          nullable: true
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time